# backend/api/test_views.py
import pytest
from django.test import TestCase
from django.contrib.auth.models import User
from django.urls import reverse
from rest_framework.test import APITestCase, APIClient
from rest_framework import status
from datetime import time
from unittest.mock import patch, Mock

from tutoring.models import TutoringStudent, Group, Lesson, Attendance, Parent
from stripeInt.models import StripeProd
from .views import LessonRollViewSet, MyTokenObtainPairView, GroupViewSet


class LessonRollViewSetTest(APITestCase):
    def setUp(self):
        # Create test user
        self.user = User.objects.create_user(
            username='testuser',
            password='testpass123'
        )
        self.client.force_authenticate(user=self.user)
        
        # Create test data
        self.parent = Parent.objects.create(name='Test Parent', stripeId='stripe_123')
        self.student1 = TutoringStudent.objects.create(parent=self.parent)
        self.student2 = TutoringStudent.objects.create(parent=self.parent)
        
        self.group = Group.objects.create(
            tutor='Test Tutor',
            course=Group.CourseChoices.JUNIOR_MATHS,
            day_of_week=Group.Weekday.MONDAY,
            time_of_day=time(10, 0)
        )
        self.group.tutoringStudents.add(self.student1, self.student2)
        
        self.lesson = Lesson.objects.create(
            group=self.group,
            notes='Test lesson notes'
        )
        
        self.attendance = Attendance.objects.create(
            lesson=self.lesson,
            tutoringStudent=self.student1,
            homework=True,
            paid=False
        )

    def test_get_roll_data(self):
        """Test GET /api/lessons/{lesson_id}/roll/"""
        url = f'/api/lessons/{self.lesson.id}/roll/'
        response = self.client.get(url)
        
        assert response.status_code == status.HTTP_200_OK
        data = response.json()
        assert data['id'] == self.lesson.id
        assert data['notes'] == 'Test lesson notes'
        assert len(data['attendances']) == 1
        assert len(data['all_students']) == 2

    def test_reset_roll(self):
        """Test DELETE /api/lessons/{lesson_id}/roll/reset/"""
        url = f'/api/lessons/{self.lesson.id}/roll/reset/'
        
        # Verify attendance exists
        assert self.lesson.attendances.count() == 1
        
        response = self.client.delete(url)
        
        assert response.status_code == status.HTTP_200_OK
        assert 'Reset 1 attendance records' in response.json()['message']
        
        # Verify attendance was deleted
        assert self.lesson.attendances.count() == 0

    def test_roll_summary(self):
        """Test GET /api/lessons/{lesson_id}/roll/summary/"""
        url = f'/api/lessons/{self.lesson.id}/roll/summary/'
        
        response = self.client.get(url)
        
        assert response.status_code == status.HTTP_200_OK
        data = response.json()
        
        assert data['lesson_id'] == self.lesson.id
        assert data['total_students_in_group'] == 2
        assert data['present_count'] == 1
        assert data['absent_count'] == 1
        assert data['homework_completed'] == 1
        assert data['homework_not_completed'] == 0
        assert data['paid_count'] == 0
        assert data['unpaid_count'] == 1
        assert data['attendance_rate'] == 50.0

    def test_roll_summary_no_group(self):
        """Test roll summary for lesson without group"""
        lesson_no_group = Lesson.objects.create(notes='No group lesson')
        url = f'/api/lessons/{lesson_no_group.id}/roll/summary/'
        
        response = self.client.get(url)
        
        assert response.status_code == status.HTTP_200_OK
        data = response.json()
        assert data['attendance_rate'] == 0

    def test_lessons_by_group(self):
        """Test GET /api/lessons/group/{group_id}/"""
        # Create another lesson for the same group
        lesson2 = Lesson.objects.create(group=self.group, notes='Second lesson')
        
        url = f'/api/lessons/group/{self.group.id}/'
        response = self.client.get(url)
        
        assert response.status_code == status.HTTP_200_OK
        data = response.json()
        
        assert data['group_info']['id'] == self.group.id
        assert data['group_info']['course'] == Group.CourseChoices.JUNIOR_MATHS
        assert data['group_info']['tutor'] == 'Test Tutor'
        assert data['total_lessons'] == 2
        assert len(data['lessons']) == 2

    def test_lessons_by_group_not_found(self):
        """Test GET with non-existent group ID"""
        url = '/api/lessons/group/9999/'
        response = self.client.get(url)
        
        assert response.status_code == status.HTTP_404_NOT_FOUND
        assert 'Group with id 9999 not found' in response.json()['error']

    def test_get_serializer_class(self):
        """Test serializer class selection based on action"""
        viewset = LessonRollViewSet()
        
        # Test roll action
        viewset.action = 'roll'
        from api.serializers import LessonRollReadSerializer
        assert viewset.get_serializer_class() == LessonRollReadSerializer
        
        # Test update_roll action
        viewset.action = 'update_roll'
        from api.serializers import LessonRollUpdateSerializer
        assert viewset.get_serializer_class() == LessonRollUpdateSerializer
        
        # Test default
        viewset.action = 'list'
        assert viewset.get_serializer_class() == LessonRollReadSerializer

    def test_unauthenticated_access(self):
        """Test that unauthenticated requests are rejected"""
        self.client.logout()
        url = f'/api/lessons/{self.lesson.id}/roll/'
        
        response = self.client.get(url)
        assert response.status_code == status.HTTP_401_UNAUTHORIZED


class MyTokenObtainPairViewTest(APITestCase):
    def setUp(self):
        self.user = User.objects.create_user(
            username='testuser',
            password='testpass123'
        )

    @patch('api.serializers.MyTokenObtainPairSerializer.validate')
    def test_token_obtain_success(self, mock_validate):
        """Test successful token obtain"""
        mock_validate.return_value = {
            'access': 'mock_access_token',
            'refresh': 'mock_refresh_token',
            'roles': []
        }
        
        url = '/api/token/'
        data = {
            'username': 'testuser',
            'password': 'testpass123'
        }
        
        response = self.client.post(url, data, format='json')
        
        # Note: This would need proper JWT setup to test fully
        # For basic coverage, we're testing the view structure
        assert hasattr(MyTokenObtainPairView, 'serializer_class')

    def test_token_obtain_invalid_credentials(self):
        """Test token obtain with invalid credentials"""
        url = '/api/token/'
        data = {
            'username': 'testuser',
            'password': 'wrongpassword'
        }
        
        response = self.client.post(url, data, format='json')
        # Would typically return 401, but depends on JWT setup
        # This ensures the view can handle bad requests


class GroupViewSetTest(APITestCase):
    def setUp(self):
        self.user = User.objects.create_user(
            username='testuser',
            password='testpass123'
        )
        self.client.force_authenticate(user=self.user)
        
        self.stripe_prod = StripeProd.objects.create(name='Test Product')
        self.group = Group.objects.create(
            lesson_length=2,
            associated_product=self.stripe_prod,
            tutor='Test Tutor',
            course=Group.CourseChoices.YEAR11_ADV,
            day_of_week=Group.Weekday.WEDNESDAY,
            time_of_day=time(14, 30)
        )

    def test_list_groups(self):
        """Test GET /api/groups/"""
        url = '/api/groups/'
        response = self.client.get(url)
        
        assert response.status_code == status.HTTP_200_OK
        data = response.json()
        assert len(data) == 1
        assert data[0]['id'] == self.group.id
        assert data[0]['tutor'] == 'Test Tutor'

    def test_retrieve_group(self):
        """Test GET /api/groups/{id}/"""
        url = f'/api/groups/{self.group.id}/'
        response = self.client.get(url)
        
        assert response.status_code == status.HTTP_200_OK
        data = response.json()
        assert data['id'] == self.group.id
        assert data['lesson_length'] == 2
        assert data['course'] == Group.CourseChoices.YEAR11_ADV
        assert data['weekly_time'] == 'Wednesday 02:30 PM'

    def test_create_group(self):
        """Test POST /api/groups/"""
        url = '/api/groups/'
        data = {
            'lesson_length': 1,
            'tutor': 'New Tutor',
            'course': Group.CourseChoices.JUNIOR_MATHS,
            'day_of_week': Group.Weekday.FRIDAY,
            'time_of_day': '16:00:00'
        }
        
        response = self.client.post(url, data, format='json')
        
        assert response.status_code == status.HTTP_201_CREATED
        assert Group.objects.count() == 2
        
        new_group = Group.objects.get(tutor='New Tutor')
        assert new_group.course == Group.CourseChoices.JUNIOR_MATHS

    def test_update_group(self):
        """Test PUT /api/groups/{id}/"""
        url = f'/api/groups/{self.group.id}/'
        data = {
            'lesson_length': 3,
            'tutor': 'Updated Tutor',
            'course': Group.CourseChoices.YEAR12_ADV,
            'day_of_week': Group.Weekday.THURSDAY,
            'time_of_day': '15:00:00'
        }
        
        response = self.client.put(url, data, format='json')
        
        assert response.status_code == status.HTTP_200_OK
        
        self.group.refresh_from_db()
        assert self.group.lesson_length == 3
        assert self.group.tutor == 'Updated Tutor'

    def test_partial_update_group(self):
        """Test PATCH /api/groups/{id}/"""
        url = f'/api/groups/{self.group.id}/'
        data = {
            'tutor': 'Partially Updated Tutor'
        }
        
        response = self.client.patch(url, data, format='json')
        
        assert response.status_code == status.HTTP_200_OK
        
        self.group.refresh_from_db()
        assert self.group.tutor == 'Partially Updated Tutor'
        # Other fields should remain unchanged
        assert self.group.lesson_length == 2

    def test_delete_group(self):
        """Test DELETE /api/groups/{id}/"""
        url = f'/api/groups/{self.group.id}/'
        
        response = self.client.delete(url)
        
        assert response.status_code == status.HTTP_204_NO_CONTENT
        assert Group.objects.count() == 0

    def test_group_not_found(self):
        """Test GET with non-existent group ID"""
        url = '/api/groups/9999/'
        response = self.client.get(url)
        
        assert response.status_code == status.HTTP_404_NOT_FOUND